+++ src/client/cypress/e2e/filters/filter.cy.ts
@@ -1,0 +1,241 @@
+import { WorkflowStatus } from "@swissgeol/ui-core";
+import { hasPagination, showTableAndWaitForData, verifyPaginationText } from "../helpers/dataGridHelpers";
+import { evaluateSelect, setInput, setSelect, setYesNoSelect } from "../helpers/formHelpers";
+import { createBorehole, goToRouteAndAcceptTerms, returnToOverview } from "../helpers/testHelpers";
+
+describe("Search filter tests", () => {
+  it("has search filters", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Filters");
+  });
+
+  // TODO: Re-add once https://github.com/swisstopo/swissgeol-boreholes-suite/issues/2380 is fixed
+  it.skip("checks that the registration filter settings control the filter visibility.", () => {
+    // precondition filters not visible
+    goToRouteAndAcceptTerms("/");
+    cy.get('[data-cy="show-filter-button"]').click();
+    cy.contains("Registration").click();
+    cy.contains("Show all fields");
+    cy.contains("Created by").should("not.exist");
+    cy.contains("Creation date").should("not.exist");
+
+    // turn on registration filters
+    cy.dataCy("settings-button").click();
+    cy.dataCy("general-tab").click();
+    cy.contains("Registration filters").click();
+    cy.contains("Select all").click();
+    cy.wait("@setting");
+
+    // check visibility of filters
+    returnToOverview();
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Registration").click();
+    cy.contains("Created by");
+    cy.contains("Creation date");
+
+    // reset setting
+    cy.dataCy("settings-button").click();
+    cy.dataCy("general-tab").click();
+    cy.contains("Registration filters").click();
+    cy.contains("Unselect all").click();
+    cy.wait("@setting");
+  });
+
+  it("filters boreholes by creator name", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Registration").click();
+    cy.dataCy("show-all-fields-switch").click();
+
+    // input value
+    setInput("created_by", "v_ U%r");
+    cy.wait("@edit_list");
+
+    // check content of table
+    showTableAndWaitForData();
+    verifyPaginationText("1–100 of 602");
+  });
+
+  it("filters boreholes national_interest and striae", () => {
+    // TODO: https://github.com/swisstopo/swissgeol-boreholes-suite/issues/2547
+    // Reactivate all commented code once lithology filter is fixed
+    createBorehole({
+      originalName: "Borehole 1 with striae: true",
+      name: "striae true / null",
+      nationalInterest: false,
+    }).as("borehole_id");
+    // .then(boreholeId => {
+    //   createStratigraphy(boreholeId, 3000)
+    //     .as("stratigraphy_id")
+    //     .then(id => {
+    //       createLithologyLayer(id, { isStriae: null });
+    //       createLithologyLayer(id, { isStriae: true });
+    //     });
+    // });
+
+    createBorehole({
+      originalName: "Borehole 1 with striae: false",
+      name: "striae false / null",
+      nationalInterest: false,
+    }).as("borehole_id2");
+    // .then(boreholeId2 => {
+    //   createStratigraphy(boreholeId2, 3000)
+    //     .as("stratigraphy_id")
+    //     .then(id => {
+    //       createLithologyLayer(id, { isStriae: false });
+    //       createLithologyLayer(id, { isStriae: null });
+    //     });
+    // });
+
+    createBorehole({
+      originalName: "Borehole 2 with striae: false",
+      name: "striae false / null",
+      nationalInterest: false,
+    }).as("borehole_id3");
+    // .then(boreholeId3 => {
+    //   createStratigraphy(boreholeId3, 3000)
+    //     .as("stratigraphy_id")
+    //     .then(id => {
+    //       createLithologyLayer(id, { isStriae: false });
+    //     });
+    // });
+
+    createBorehole({
+      originalName: "Borehole 3 with striae: false",
+      name: "striae false, national_interest null",
+      nationalInterest: null,
+    }).as("borehole_id3");
+    // .then(boreholeId3 => {
+    //   createStratigraphy(boreholeId3, 3000)
+    //     .as("stratigraphy_id")
+    //     .then(id => {
+    //       createLithologyLayer(id, { isStriae: false });
+    //     });
+    // });
+
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Location").click();
+    cy.dataCy("show-all-fields-switch").click();
+    setYesNoSelect("national_interest", "Yes");
+    cy.wait("@edit_list");
+
+    showTableAndWaitForData();
+    verifyPaginationText("1–100 of 300");
+    cy.get('[data-cy="filter-chip-national_interest"]').should("exist");
+
+    setYesNoSelect("national_interest", "Not specified");
+    cy.wait("@edit_list");
+    hasPagination(false);
+    cy.get('[data-cy="filter-chip-national_interest"]').should("exist");
+
+    setYesNoSelect("national_interest", "No");
+    cy.wait("@edit_list");
+    verifyPaginationText("1–100 of 2703");
+    cy.get('[data-cy="filter-chip-national_interest"]').should("exist");
+
+    // reset national interest filter
+    cy.get('[data-cy="filter-chip-national_interest"]')
+      .should("exist")
+      .within(() => {
+        cy.get("svg").click();
+      });
+
+    cy.wait("@edit_list");
+    cy.get('[data-cy="filter-chip-national_interest"]').should("not.exist");
+    verifyPaginationText("1–100 of 3004");
+  });
+
+  it("filters boreholes by status", () => {
+    goToRouteAndAcceptTerms("");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Status").click();
+    cy.dataCy("boreholes-number-preview").should("have.text", "3'000");
+    cy.dataCy(WorkflowStatus.Draft).click();
+    cy.dataCy("boreholes-number-preview").should("have.text", "3'000");
+    cy.dataCy(WorkflowStatus.Reviewed).click();
+    cy.dataCy("boreholes-number-preview").should("have.text", "0");
+  });
+
+  it("filters boreholes by boreholestatus", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Borehole").click();
+    cy.dataCy("show-all-fields-switch").click();
+    setSelect("status", 2);
+    cy.wait("@edit_list");
+
+    cy.get('[data-cy="boreholes-number-preview"]').should("have.text", "314");
+    cy.get('[data-cy="filter-chip-boreholestatus"]').contains("Borehole status");
+  });
+
+  // Todo: readd when new filtes are implemented https://github.com/swisstopo/swissgeol-boreholes-suite/issues/2547
+  it.skip("filters boreholes by color and uscs3", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Lithology").click();
+    cy.dataCy("show-all-fields-switch").click();
+    setSelect("color", 0);
+    cy.wait("@edit_list");
+    showTableAndWaitForData();
+    verifyPaginationText("1–100 of 1403");
+    setSelect("uscs_3", 4);
+    cy.wait("@edit_list");
+
+    // check content of table
+    verifyPaginationText("1–100 of 196");
+    cy.get(".MuiDataGrid-row").contains("Aisha Thiel").should("exist");
+  });
+
+  // Todo: readd when new filtes are implemented https://github.com/swisstopo/swissgeol-boreholes-suite/issues/2547
+  it.skip("filters boreholes by original lithology in editor mode", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Lithology").click();
+    cy.dataCy("show-all-fields-switch").click();
+    setInput("original_lithology", "Wooden Chair");
+    cy.wait("@edit_list");
+    showTableAndWaitForData();
+    hasPagination(false);
+  });
+
+  it("filters boreholes by creation date", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Registration").click();
+    cy.dataCy("show-all-fields-switch").click();
+
+    setInput("created_date_from", "2021-11-09");
+    cy.wait("@edit_list");
+    setInput("created_date_to", "2021-11-10");
+
+    cy.wait("@edit_list");
+
+    // check content of table
+    showTableAndWaitForData();
+    hasPagination(false);
+  });
+
+  it("filters boreholes by workgroup", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    showTableAndWaitForData();
+    cy.contains("Workgroup").click();
+    cy.wait("@borehole");
+    cy.dataCy("filter-chip-workgroup").should("not.exist");
+    cy.dataCy("Default").click();
+    cy.dataCy("filter-chip-workgroup").should("exist");
+    cy.dataCy("all").click();
+    cy.dataCy("filter-chip-workgroup").should("not.exist");
+  });
+
+  it("shows additional values in domain filter", () => {
+    goToRouteAndAcceptTerms("/");
+    cy.dataCy("show-filter-button").click();
+    cy.contains("Location").click();
+    cy.dataCy("show-all-fields-switch").click();
+    setSelect("borehole_identifier", 0);
+    evaluateSelect("borehole_identifier", "Boreholes.swissgeol.ch ID");
+  });
+});